+++ src/components/NavBar.vue
@@ -4,6 +4,11 @@
       <router-link style="margin-right:3%" class="navbar-brand" to="/">Project-Manager</router-link>
 <!--      <button @click="toggleSidebar" class="btn btn-dark" style="padding-right:2%"><i class="bi bi-chat-right-text-fill"></i></button>-->
       <span class="white-text">{{keycloak.getEmail()}}</span>
+      <div>
+        <button @click="logout" class="btn btn-outline-danger">
+          <i class="bi bi-box-arrow-right"></i> logout
+        </button>
+      </div>
     </nav>
   </div>
 </template>
@@ -22,6 +27,9 @@ export default {
     toggleSidebar() {
       this.$store.commit('toggleSidebar');
     },
+    logout(){
+      keycloak.logout();
+    }
   },
 };
 </script>