+++ CHANGELOG.md
@@ -44,3 +44,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fetch visible bridgeheads and fetch project bridgeheads
 - Refresh Keycloak token
 - Redirect to explorer Url for editing bridgeheads or query
+- Keycloak logout